Fixes and Transfer, Deposit, and Extract

+++ ProjectBank/src/main/java/com/franbuss/ProjectBank/services/serviceImpl/UserServiceImpl.java
@@ -43,6 +43,18 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
+    public UserResponseDTO createEmployee(UserRegisterRequestDTO userRegisterRequestDTO) throws Exception {
+        Optional<User> user = userRepository.findByDni(userRegisterRequestDTO.getDni());
+        if (user.isPresent()) {
+            throw new Exception("User already exists");
+        }
+        User userMapped = modelMapper.map(userRegisterRequestDTO, User.class);
+        userMapped.setRol(Rol.EMPLEADO);
+        User saveUser = userRepository.save(userMapped);
+        return modelMapper.map(saveUser, UserResponseDTO.class);
+    }
+
+    @Override
     public UserResponseDTO updateUser(Long id, UserUpdateRequestDTO userUpdateRequestDTO) throws Exception {
         Optional<User> user = userRepository.findById(id);
         if (user.isPresent()) {
@@ -64,8 +76,28 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+    public void deleteUser(Long id) throws Exception {
+        Optional<User> user = userRepository.findById(id);
+        if (user.isPresent()) {
+            User userToBeDelete = user.get();
+            if (userToBeDelete.getCheckOut()){
+                userRepository.deleteById(id);
+            }
+        } else {
+            throw new Exception("User cannot be deleted");
+        }
+    }
+
+    @Override
+    public void checkOut(Long id) throws Exception {
+        Optional<User> user = userRepository.findById(id);
+        if (user.isPresent()) {
+            User optionalUser = user.get();
+            optionalUser.setCheckOut(true);
+            userRepository.save(optionalUser);
+        } else {
+            throw new Exception("User not found");
+        }
     }
 
     @Override
